split up tests with downloading and without downloading
